localagent/requirements.txt
requests
python-socketio
python-nmap
msgpack